iemInitDecoderAndPrefetchOpcodes: Outside 64-bit mode, effective address calculation shall be done using 32-bit precision according to 4.5.3, AMD64 Vol. 2. See xTracker #7408.
svn:sync-xref-src-repo-rev: r95494

+++ src/VBox/VMM/VMMAll/IEMAll.cpp
@@ -1,4 +1,4 @@
-/* $Id: IEMAll.cpp 52361 2014-08-12 19:04:10Z knut.osmundsen@oracle.com $ */
+/* $Id: IEMAll.cpp 52362 2014-08-12 19:04:53Z knut.osmundsen@oracle.com $ */
 /** @file
  * IEM - Interpreted Execution Manager - All Contexts.
  */
@@ -1116,7 +1116,7 @@ static VBOXSTRICTRC iemOpcodeFetchMoreBytes(PIEMCPU pIemCpu, size_t cbMin)
         }
         if (cbToTryRead < cbMin - cbLeft)
             return iemRaiseSelectorBounds(pIemCpu, X86_SREG_CS, IEM_ACCESS_INSTRUCTION);
-        GCPtrNext = pCtx->cs.u64Base + GCPtrNext32;
+        GCPtrNext = (uint32_t)pCtx->cs.u64Base + GCPtrNext32;
     }
 
     /* Only read up to the end of the page, and make sure we don't read more
